requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
<<<<<<< HEAD
gunicorn==21.2.0
bcrypt==3.2.0
sqlalchemy==2.0.29
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
jinja2==3.1.4
google-generativeai>=0.8.0
google-cloud-speech==2.26.1
google-cloud-texttospeech==2.16.3
psycopg2-binary
passlib[argon2]
argon2-cffi
=======
python-dotenv==1.0.0
google-genai==0.3.0
sqlalchemy==2.0.29
python-multipart==0.0.9
google-cloud-texttospeech==2.16.3
psycopg2-binary
>>>>>>> b72a9f6 (backend)
